Create scraper.py
A programmer that takes the https://www.nature.com/nature/articles?sort=PubDate&year=2020&page=3 URL and then scans the page source for articles.
Determine the article type and the ending page number to view all articles of that type.
The content of each article of the "News" type is saved, that is, the text from the body of the article without tags, in a separate file named% article_title% .txt. When you save the file, spaces in the article title are replaced with underscores and punctuation marks in the file name are removed. For example, an article titled "The legendary Arecibo telescope will close forever - scientists stagger" should be saved to a file named Legendary_Arecibo_telescope_will_close_forever_scientists_are_reeling.txt.